+++ src/NEWS.md
@@ -11,6 +11,8 @@
 ### bug fixes
 
 - double showing of progress-bar component
+- fixed scrollbar issues (now always hidden)
+- fixed bundle list side bar animation
 
 ## Version 1.3.2